♻️ refactor: 이메일 모듈 분리
