+++ api/src/main/java/com/sponus/sponusbe/domain/organization/service/OrganizationService.java
@@ -3,9 +3,7 @@ package com.sponus.sponusbe.domain.organization.service;
 import static com.sponus.sponusbe.domain.organization.exception.OrganizationErrorCode.*;
 
 import java.util.List;
-import java.util.Random;
 
-import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,6 +13,7 @@ import com.sponus.coredomain.domain.notification.Notification;
 import com.sponus.coredomain.domain.notification.repository.NotificationRepository;
 import com.sponus.coredomain.domain.organization.Organization;
 import com.sponus.coredomain.domain.organization.repository.OrganizationRepository;
+import com.sponus.coreinfraemail.EmailUtil;
 import com.sponus.coreinfras3.S3Util;
 import com.sponus.sponusbe.domain.notification.exception.NotificationErrorCode;
 import com.sponus.sponusbe.domain.notification.exception.NotificationException;
@@ -24,8 +23,6 @@ import com.sponus.sponusbe.domain.organization.dto.OrganizationSummaryResponse;
 import com.sponus.sponusbe.domain.organization.dto.OrganizationUpdateRequest;
 import com.sponus.sponusbe.domain.organization.exception.OrganizationException;
 
-import jakarta.mail.internet.InternetAddress;
-import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -38,8 +35,8 @@ public class OrganizationService {
 	private final OrganizationRepository organizationRepository;
 	private final NotificationRepository notificationRepository;
 	private final PasswordEncoder passwordEncoder;
-	private final JavaMailSender emailSender;
 	private final S3Util s3Util;
+	private final EmailUtil emailUtil;
 
 	public OrganizationJoinResponse join(OrganizationJoinRequest request) {
 		final Organization organization = organizationRepository.save(
@@ -72,57 +69,7 @@ public class OrganizationService {
 	}
 
 	public String sendEmail(String to) throws Exception {
-		String code = createEmailCode();
-		MimeMessage message = createEmail(to, code);
-		emailSender.send(message);
-		return code;
-	}
-
-	private MimeMessage createEmail(String to, String code) throws Exception {
-		log.info("보내는 대상 : " + to);
-		log.info("인증 코드 : " + code);
-		MimeMessage message = emailSender.createMimeMessage();
-		message.addRecipients(MimeMessage.RecipientType.TO, to); // 보내는 대상
-		message.setSubject("Spon-us 인증 코드 발급입니다."); // 제목
-
-		String msgg = "";
-		msgg += "<div style='margin:20px;'>";
-		msgg += "<h1> Spon-us 인증 코드입니다. </h1>"; // 동적으로 메시지 내용 변경
-		msgg += "<br>";
-		msgg += "<div align='center' style='border:1px solid black; font-family:verdana';>";
-		msgg += "<h3 style='color:blue;'> 인증 코드 </h3>";
-		msgg += "<div style='font-size:130%'>";
-		msgg += "인증 코드 : <strong>";
-		msgg += code + "</strong><div><br/> ";
-		msgg += "</div>";
-		message.setText(msgg, "utf-8", "html"); // 내용
-		message.setFrom(new InternetAddress("Spon-us_team", "Spon-us")); // 보내는 사람
-
-		return message;
-	}
-
-	public static String createEmailCode() {
-		StringBuilder code = new StringBuilder();
-		Random rnd = new Random();
-
-		for (int i = 0; i < 6; i++) { // 인증 코드 6자리
-			int index = rnd.nextInt(3); // 0~2 까지 랜덤
-			switch (index) {
-				case 0:
-					code.append((char)((rnd.nextInt(26)) + 97));
-					//  a~z  (ex. 1+97=98 => (char)98 = 'b')
-					break;
-				case 1:
-					code.append((char)((rnd.nextInt(26)) + 65));
-					//  A~Z
-					break;
-				default:
-					code.append((rnd.nextInt(10)));
-					// 0~9
-					break;
-			}
-		}
-		return code.toString();
+		return emailUtil.sendEmail(to);
 	}
 
 	public List<OrganizationSummaryResponse> searchOrganization(String keyword) {
