+++ core/core-infra-email/build.gradle
@@ -1,0 +1,7 @@
+dependencies {
+    // email
+    implementation 'org.springframework.boot:spring-boot-starter-mail'
+}
+
+bootJar { enabled = false }
+jar { enabled = true }
