+++ settings.gradle
@@ -4,3 +4,4 @@ include 'core:core-domain'
 include 'core:core-infra-db'
 include 'core:core-infra-s3'
 include 'core:core-infra-redis'
+include 'core:core-infra-email'
