+++ src/main/resources/application.yml
@@ -1,0 +1,9 @@
+spring:
+  profiles:
+    active: local
+    include: secret
+
+  servlet:
+    multipart:
+      max-file-size: 20MB
+      max-request-size: 20MB
